Add folder structure and Router
Added the folders assets for images and other assets, components for different react components, pages for the pages (currenty only one page is planned for the website), services for services that might be used many times, for example the websocket client. Further i added a Router in the app.js that should be used to route the visitors to the home page (and other pages if so necessary).